various bugfixes and brought test cases back to speed
