+++ lib/Intersection.js
@@ -1,4 +1,3 @@
-
 /**
  *  Intersection
  */
