+++ lib/IntersectionParams.js
@@ -100,17 +100,17 @@ IntersectionParams.prototype.init = function (type, params) {
 
 IntersectionParams.TYPE = {};
 var IPTYPE = IntersectionParams.TYPE;
-IPTYPE.LINE = { value: 'Line' };
-IPTYPE.RECT = { value: 'Rectangle' };
-IPTYPE.ROUNDRECT = { value: 'RoundRectangle' };
-IPTYPE.CIRCLE = { value: 'Circle' };
-IPTYPE.ELLIPSE = { value: 'Ellipse' };
-IPTYPE.POLYGON = { value: 'Polygon' };
-IPTYPE.POLYLINE = { value: 'Polyline' };
-IPTYPE.PATH = { value: 'Path' };
-IPTYPE.ARC = { value: 'Arc' };
-IPTYPE.BEZIER2 = { value: 'Bezier2' };
-IPTYPE.BEZIER3 = { value: 'Bezier3' });
+IPTYPE.LINE = 'Line';
+IPTYPE.RECT = 'Rectangle';
+IPTYPE.ROUNDRECT = 'RoundRectangle';
+IPTYPE.CIRCLE = 'Circle';
+IPTYPE.ELLIPSE = 'Ellipse';
+IPTYPE.POLYGON = 'Polygon';
+IPTYPE.POLYLINE = 'Polyline';
+IPTYPE.PATH = 'Path';
+IPTYPE.ARC = 'Arc';
+IPTYPE.BEZIER2 = 'Bezier2';
+IPTYPE.BEZIER3 = 'Bezier3';
 
 
 function parsePointsString(points) {
@@ -120,7 +120,7 @@ function parsePointsString(points) {
     });
 }
 
-IntersectionParams.newShape(type, props) {
+IntersectionParams.newShape = function(type, props) {
     if(type === IPTYPE.RECT && (props.rx > 0 || props.ry > 0)) {
         type = IPTYPE.ROUNDRECT;
     }
