+++ lib/functions/bezier.js
@@ -1,4 +1,4 @@
-var Intersection = require('./Intersection');
+var Intersection = require('../Intersection');
 
 var affine = require('kld-affine');
 var Point2D = affine.Point2D;
