+++ lib/intersect.js
@@ -19,44 +19,140 @@ var Intersection = require('./Intersection');
 var IPTYPE = IntersectionParams.TYPE;
 
 
+
+/**
+ *  bezout
+ *
+ *  This code is based on MgcIntr2DElpElp.cpp written by David Eberly.  His
+ *  code along with many other excellent examples are avaiable at his site:
+ *  http://www.geometrictools.com
+ *
+ *  @param {Array<Point2D>} e1
+ *  @param {Array<Point2D>} e2
+ *  @returns {Polynomial}
+ */
+function bezout(e1, e2) {
+    var AB    = e1[0]*e2[1] - e2[0]*e1[1];
+    var AC    = e1[0]*e2[2] - e2[0]*e1[2];
+    var AD    = e1[0]*e2[3] - e2[0]*e1[3];
+    var AE    = e1[0]*e2[4] - e2[0]*e1[4];
+    var AF    = e1[0]*e2[5] - e2[0]*e1[5];
+    var BC    = e1[1]*e2[2] - e2[1]*e1[2];
+    var BE    = e1[1]*e2[4] - e2[1]*e1[4];
+    var BF    = e1[1]*e2[5] - e2[1]*e1[5];
+    var CD    = e1[2]*e2[3] - e2[2]*e1[3];
+    var DE    = e1[3]*e2[4] - e2[3]*e1[4];
+    var DF    = e1[3]*e2[5] - e2[3]*e1[5];
+    var BFpDE = BF + DE;
+    var BEmCD = BE - CD;
+
+    return new Polynomial(
+        AB*BC - AC*AC,
+        AB*BEmCD + AD*BC - 2*AC*AE,
+        AB*BFpDE + AD*BEmCD - AE*AE - 2*AC*AF,
+        AB*DF + AD*BFpDE - 2*AE*AF,
+        AD*DF - AF*AF
+    );
+}
+
+/**
+    Removes from intersection points those points that are not between two rays determined by arc parameters.
+    Rays begin at ellipse center and go through arc startPoint/endPoint.
+
+    @param {Intersection} intersection - will be modified and returned
+    @param {Point2D} c - center of arc ellipse
+    @param {Number} rx
+    @param {Number} ry
+    @param {Number} phi - in radians
+    @param {Number} th1 - in radians
+    @param {Number} dth - in radians
+    @param {Matrix2D} [m] - arc transformation matrix
+    @returns {Intersection}
+*/
+function removePointsNotInArc(intersection, c, rx, ry, phi, th1, dth, m) {
+    if (intersection.points.length === 0) return intersection;
+    if (m && !m.isIdentity())
+        var mp = m.inverse();
+    var np = [];
+    var vx = new Vector2D(1, 0);
+    var pi2 = Math.PI * 2;
+    var wasNeg = dth < 0;
+    var wasBig = Math.abs(dth) > Math.PI;
+    var m1 = new Matrix2D().scaleNonUniform(1, ry / rx).rotate(th1);
+    var m2 = new Matrix2D().scaleNonUniform(1, ry / rx).rotate(th1 + dth);
+
+    th1 = (vx.angleBetween(vx.transform(m1)) + pi2) % pi2;
+    dth = vx.transform(m1).angleBetween(vx.transform(m2));
+    dth = (wasBig ? pi2 - Math.abs(dth) : Math.abs(dth)) * (wasNeg ? -1 : 1);
+    var m3 = new Matrix2D().rotate(phi).multiply(m1);
+
+    for (var i = 0, p, a; i < intersection.points.length; i++) {
+        p = intersection.points[i];
+        a = vx.transform(m3).angleBetween(Vector2D.fromPoints(c, (mp) ? p.transform(mp) : p));
+        if (dth >= 0) {
+            a = (a + 2 * pi2) % pi2;
+            if (a <= dth)
+                np.push(p);
+        } else {
+            a = (a - 2 * pi2) % pi2;
+            if (a >= dth)
+                np.push(p);
+        }
+    }
+    intersection.points = np;
+    return intersection;
+};
+
+/**
+    points1 will be modified, points close (almost identical) to any point in points2 will be removed
+
+    @param {Array<Point2D>} points1 - will be modified, points close to any point in points2 will be removed
+    @param {Array<Point2D>} points2
+*/
+function removeClosePoints(points1, points2) {
+    if (points1.length === 0 || points2.length === 0)
+        return;
+    var maxf = function (p, v) { if (p < v.x) p = v.x; if (p < v.y) p = v.y; return p; };
+    var max = points1.reduce(maxf, 0);
+    max = points2.reduce(maxf, max);
+    var ERRF = 1e-15;
+    var ZEROepsilon = 100 * max * ERRF * Math.SQRT2;
+    var j;
+    for (var i = 0; i < points1.length;) {
+        for (j = 0; j < points2.length; j++) {
+            if (points1[i].distanceFrom(points2[j]) <= ZEROepsilon) {
+                points1.splice(i, 1);
+                break;
+            }
+        }
+        if (j == points2.length)
+            i++;
+    }
+}
+
+
+
+
 module.exports = function() {
     var argLength = arguments.length;
     if(argLength === 0) {
         throw new Error("The intersect module expects intersection functions to be provided for dependency injection");
     }
 
-    var intersectionFunctions;
 
-    for(var i = 0; i < argLength; i++) {
-        var arg = arguments[i];
-        for(var key in arg) {
-            if(arg.hasOwnProperty(key)) {
-                intersectionFunctions[key] = arg[key];
-            }
-        }
-    }
-
-    var composedShapeMethods = {};
-    composedShapeMethods[IPTYPE.PATH] = intersectionFunctions.intersectPathShape;
-    composedShapeMethods[IPTYPE.POLYLINE] = intersectionFunctions.intersectLinesShape;
-    composedShapeMethods[IPTYPE.POLYGON] = intersectionFunctions.intersectLinesShape;
-    composedShapeMethods[IPTYPE.RECT] = intersectionFunctions.intersectLinesShape;
-    composedShapeMethods[IPTYPE.ROUNDRECT] = intersectionFunctions.intersectPathShape;
-    composedShapeMethods[IPTYPE.ARC] = intersectionFunctions.intersectArcShape;
-
-
-    return function intersect(shape1, shape2, m1, m2) {
+    function intersect(shape1, shape2, m1, m2) {
         var ip1 = shape1;
         var ip2 = shape2;
         var result;
+console.log("I", ip1.type, ip2.type);
 
         if (ip1 !== null && ip2 !== null) {
             var method;
 
-            if (method = stat.composedShapeMethods[ip1.type]) {
+            if (method = composedShapeMethods[ip1.type]) {
                 result = method(ip1, ip2, m1, m2);
             }
-            else if (method = stat.composedShapeMethods[ip2.type]) {
+            else if (method = composedShapeMethods[ip2.type]) {
                 result = method(ip2, ip1, m2, m1);
             }
             else {
@@ -174,7 +270,7 @@ module.exports = function() {
                         }
                     }
                 }
-
+console.log(type1, type2)
                 if (type1 < type2) {
                     method = "intersect" + type1 + type2;
                     params = params1.concat(params2);
@@ -197,5 +293,424 @@ module.exports = function() {
 
         return result;
     }
+
+
+
+
+
+    var intersectionFunctions = {
+                
+        /**
+            intersectPathShape
+
+            @param {IntersectionParams} path
+            @param {IntersectionParams} shape
+            @param {Matrix2D} [m1]
+            @param {Matrix2D} [m2]
+            @returns {Intersection}
+        */
+        intersectPathShape: function (path, shape, m1, m2) {
+            var result = new Intersection();
+            var pathParams = path.params[0];
+            var inter0;
+            var previnter;
+            console.log("IPS")
+            for (var inter, i = 0; i < pathParams.length; i++) {
+                inter = intersect(pathParams[i], shape, m1, m2);
+                if (!inter0)
+                    inter0 = inter;
+                if (previnter) {
+                    removeClosePoints(previnter.points, inter.points);
+                    result.appendPoints(previnter.points);
+                }
+                previnter = inter;
+            }
+            if (previnter) {
+                result.appendPoints(previnter.points);
+            }
+            return result;
+        },
+
+
+        /**
+            intersectLinesShape
+
+            @param {IntersectionParams} lines - IntersectionParams with points as first parameter (like types RECT, POLYLINE or POLYGON)
+            @param {IntersectionParams} shape - IntersectionParams of other shape
+            @param {Matrix2D} [m1]
+            @param {Matrix2D} [m2]
+            @param {Boolean} [closed] - if set, determines if line between first and last point will be taken into callculation too. If not set, it's true for RECT and POLYGON, false for other <b>lines</b> types.
+            @returns {Intersection}
+        */
+        intersectLinesShape: function (lines, shape, m1, m2, closed) {
+            var IPTYPE = IntersectionParams.TYPE;
+            var line_points = lines.params[0];
+            var ip = new IntersectionParams(IPTYPE.LINE, [0, 0]);
+            var result = new Intersection();
+            var inter, i;
+            var intersectLine = function (i1, i2) {
+                ip.params[0] = line_points[i1];
+                ip.params[1] = line_points[i2];
+                inter = intersect(ip, shape, m1, m2);
+                removeClosePoints(inter.points, [line_points[i2]]);
+                result.appendPoints(inter.points);
+            }
+            for (i = 0; i < line_points.length - 1; i++) {
+                intersectLine(i, i + 1);
+            }
+            if (typeof closed !== 'undefined' && closed || lines.type === IPTYPE.RECT || lines.type === IPTYPE.POLYGON) {
+                intersectLine(line_points.length - 1, 0);
+            }
+            return result;
+        },
+
+        ///////////////////////////////////////////////////////////////////
+        /**
+            intersectArcShape
+
+            @param {IntersectionParams} arc
+            @param {IntersectionParams} shape
+            @param {Matrix2D} [m1]
+            @param {Matrix2D} [m2]
+            @returns {Intersection}
+        */
+        intersectArcShape: function (arc, shape, m1, m2) {
+            m1 = m1 || Matrix2D.IDENTITY;
+            m2 = m2 || Matrix2D.IDENTITY;
+            var c1 = arc.params[0],
+                rx1 = arc.params[1],
+                ry1 = arc.params[2],
+                phi1 = arc.params[3],
+                th1 = arc.params[4],
+                dth1 = arc.params[5];
+
+            var res;
+            if (m1.isIdentity() && phi1 === 0) {
+                res = intersect(IntersectionParams.newEllipse(c1, rx1, ry1), shape, m1, m2);
+            }
+            else {
+                m1 = m1.multiply(Matrix2D.IDENTITY.translate(c1.x, c1.y).rotate(phi1));
+                c1 = new Point2D(0, 0);
+                phi1 = 0;
+                res = intersect(IntersectionParams.newEllipse(c1, rx1, ry1), shape, m1, m2);
+            }
+            res = removePointsNotInArc(res, c1, rx1, ry1, phi1, th1, dth1, m1);
+            return res;
+        },
+
+        /**
+         *  Finds intersection points of two ellipses. <br/>
+         *
+         *  This code is based on MgcIntr2DElpElp.cpp written by David Eberly. His
+         *  code along with many other excellent examples are avaiable at his site:
+         *  http://www.geometrictools.com
+         *
+         *  Changes - 2015 Robert Benko (Quazistax)
+         *
+         *  @param {Point2D} c1
+         *  @param {Number} rx1
+         *  @param {Number} ry1
+         *  @param {Point2D} c2
+         *  @param {Number} rx2
+         *  @param {Number} ry2
+         *  @returns {Intersection}
+         */
+        intersectEllipseEllipse: function (c1, rx1, ry1, c2, rx2, ry2) {
+            var a = [
+                ry1 * ry1, 0, rx1 * rx1, -2 * ry1 * ry1 * c1.x, -2 * rx1 * rx1 * c1.y,
+                ry1 * ry1 * c1.x * c1.x + rx1 * rx1 * c1.y * c1.y - rx1 * rx1 * ry1 * ry1
+            ];
+            var b = [
+                ry2 * ry2, 0, rx2 * rx2, -2 * ry2 * ry2 * c2.x, -2 * rx2 * rx2 * c2.y,
+                ry2 * ry2 * c2.x * c2.x + rx2 * rx2 * c2.y * c2.y - rx2 * rx2 * ry2 * ry2
+            ];
+
+            var yPoly = bezout(a, b);
+            var yRoots = yPoly.getRoots();
+            var epsilon = 1e-3;
+            var norm0 = (a[0] * a[0] + 2 * a[1] * a[1] + a[2] * a[2]) * epsilon;
+            var norm1 = (b[0] * b[0] + 2 * b[1] * b[1] + b[2] * b[2]) * epsilon;
+            var result = new Intersection();
+
+            var i;
+            //Handling root calculation error causing not detecting intersection
+            var clip = function (val, min, max) { return Math.max(min, Math.min(max, val)); };
+            for (i = 0 ; i < yRoots.length; i++) {
+                yRoots[i] = clip(yRoots[i], c1.y - ry1, c1.y + ry1);
+                yRoots[i] = clip(yRoots[i], c2.y - ry2, c2.y + ry2);
+            }
+
+            //For detection of multiplicated intersection points
+            yRoots.sort(function (a, b) { return a - b; });
+            var rootPointsN = [];
+
+            for (var y = 0; y < yRoots.length; y++) {
+                var xPoly = new Polynomial(
+                    a[0],
+                    a[3] + yRoots[y] * a[1],
+                    a[5] + yRoots[y] * (a[4] + yRoots[y] * a[2])
+                );
+                var ERRF = 1e-15;
+                if (Math.abs(xPoly.coefs[0]) < 10 * ERRF * Math.abs(xPoly.coefs[2]))
+                    xPoly.coefs[0] = 0;
+                var xRoots = xPoly.getRoots();
+
+                rootPointsN.push(0);
+                for (var x = 0; x < xRoots.length; x++) {
+                    var test =
+                        (a[0] * xRoots[x] + a[1] * yRoots[y] + a[3]) * xRoots[x] +
+                        (a[2] * yRoots[y] + a[4]) * yRoots[y] + a[5];
+                    if (Math.abs(test) < norm0) {
+                        test =
+                            (b[0] * xRoots[x] + b[1] * yRoots[y] + b[3]) * xRoots[x] +
+                            (b[2] * yRoots[y] + b[4]) * yRoots[y] + b[5];
+                        if (Math.abs(test) < norm1) {
+                            result.appendPoint(new Point2D(xRoots[x], yRoots[y]));
+                            rootPointsN[y] += 1;
+                        }
+                    }
+                }
+            }
+
+            if (result.points.length <= 0)
+                return result;
+
+            //Removal of multiplicated intersection points
+            var pts = result.points;
+            if (pts.length == 8) {
+                pts = pts.splice(0, 6);
+                pts.splice(2, 2);
+            }
+            else if (pts.length == 7) {
+                pts = pts.splice(0, 6);
+                pts.splice(2, 2);
+                pts.splice(rootPointsN.indexOf(1), 1);
+            }
+            else if (pts.length == 6) {
+                pts.splice(2, 2);
+                //console.log('ElEl 6pts: N: ' + rootPointsN.toString());
+                if (rootPointsN.indexOf(0) > -1) {
+                    if (pts[0].distanceFrom(pts[1]) < pts[2].distanceFrom(pts[3])) {
+                        pts.splice(0, 1);
+                    }
+                    else {
+                        pts.splice(2, 1);
+                    }
+                }
+                else if (rootPointsN[0] == rootPointsN[3]) {
+                    pts.splice(1, 2);
+                }
+            }
+            else if (pts.length == 4) {
+                if (
+                    (yRoots.length == 2)
+                || (yRoots.length == 4 && (rootPointsN[0] == 2 && rootPointsN[1] == 2 || rootPointsN[2] == 2 && rootPointsN[3] == 2))
+                ) {
+                    pts.splice(2, 2);
+                }
+            }
+            else if (pts.length == 3 || pts.length == 5) {
+                i = rootPointsN.indexOf(2);
+                if (i > -1) {
+                    if (pts.length == 3)
+                        i = i % 2;
+                    var ii = i + (i % 2 ? -1 : 2);
+                    var d1, d2, d3;
+                    d1 = pts[i].distanceFrom(pts[i + 1]);
+                    d2 = pts[i].distanceFrom(pts[ii]);
+                    d3 = pts[i + 1].distanceFrom(pts[ii]);
+                    if (d1 < d2 && d1 < d3) {
+                        pts.splice(i, 1);
+                    }
+                    else {
+                        pts.splice(ii, 1);
+                    }
+                }
+            }
+
+            var poly = yPoly;
+            var ZEROepsilon = yPoly.zeroErrorEstimate();
+            ZEROepsilon *= 100 * Math.SQRT2;
+            for (i = 0; i < pts.length - 1;) {
+                if (pts[i].distanceFrom(pts[i + 1]) < ZEROepsilon) {
+                    pts.splice(i + 1, 1);
+                    continue;
+                }
+                i++;
+            }
+
+            result.points = pts;
+            return result;
+        },
+
+
+        /**
+         *  intersectEllipseLine
+         *
+         *  NOTE: Rotation will need to be added to this function
+         *
+         *  @param {Point2D} c
+         *  @param {Number} rx
+         *  @param {Number} ry
+         *  @param {Point2D} a1
+         *  @param {Point2D} a2
+         *  @returns {Intersection}
+         */
+        intersectEllipseLine: function(c, rx, ry, a1, a2) {
+            var result;
+            var origin = new Vector2D(a1.x, a1.y);
+            var dir    = Vector2D.fromPoints(a1, a2);
+            var center = new Vector2D(c.x, c.y);
+            var diff   = origin.subtract(center);
+            var mDir   = new Vector2D( dir.x/(rx*rx),  dir.y/(ry*ry)  );
+            var mDiff  = new Vector2D( diff.x/(rx*rx), diff.y/(ry*ry) );
+
+            var a = dir.dot(mDir);
+            var b = dir.dot(mDiff);
+            var c = diff.dot(mDiff) - 1.0;
+            var d = b*b - a*c;
+
+            var ERRF = 1e-15;
+            var ZEROepsilon = 10 * Math.max(Math.abs(a), Math.abs(b), Math.abs(c)) * ERRF;
+            if (Math.abs(d) < ZEROepsilon) {
+                d = 0;
+            }
+
+            if ( d < 0 ) {
+                result = new Intersection("Outside");
+            } else if ( d > 0 ) {
+                var root = Math.sqrt(d);
+                var t_a  = (-b - root) / a;
+                var t_b  = (-b + root) / a;
+
+                t_b = (t_b > 1) ? t_b - ERRF : (t_b < 0) ? t_b + ERRF : t_b;
+                t_a = (t_a > 1) ? t_a - ERRF : (t_a < 0) ? t_a + ERRF : t_a;
+
+                if ( (t_a < 0 || 1 < t_a) && (t_b < 0 || 1 < t_b) ) {
+                    if ( (t_a < 0 && t_b < 0) || (t_a > 1 && t_b > 1) )
+                        result = new Intersection("Outside");
+                    else
+                        result = new Intersection("Inside");
+                } else {
+                    result = new Intersection();
+                    if ( 0 <= t_a && t_a <= 1 )
+                        result.appendPoint( a1.lerp(a2, t_a) );
+                    if ( 0 <= t_b && t_b <= 1 )
+                        result.appendPoint( a1.lerp(a2, t_b) );
+                }
+            } else {
+                var t = -b/a;
+                if ( 0 <= t && t <= 1 ) {
+                    result = new Intersection();
+                    result.appendPoint( a1.lerp(a2, t) );
+                } else {
+                    result = new Intersection("Outside");
+                }
+            }
+
+            return result;
+        },
+
+
+        /**
+         *  intersectLineLine
+         *
+         *  @param {Point2D} a1
+         *  @param {Point2D} a2
+         *  @param {Point2D} b1
+         *  @param {Point2D} b2
+         *  @returns {Intersection}
+         */
+        intersectLineLine: function(a1, a2, b1, b2) {
+            var result;
+
+            var ua_t = (b2.x - b1.x) * (a1.y - b1.y) - (b2.y - b1.y) * (a1.x - b1.x);
+            var ub_t = (a2.x - a1.x) * (a1.y - b1.y) - (a2.y - a1.y) * (a1.x - b1.x);
+            var u_b  = (b2.y - b1.y) * (a2.x - a1.x) - (b2.x - b1.x) * (a2.y - a1.y);
+
+            if ( u_b !== 0 ) {
+                var ua = ua_t / u_b;
+                var ub = ub_t / u_b;
+
+                if ( 0 <= ua && ua <= 1 && 0 <= ub && ub <= 1 ) {
+                    result = new Intersection();
+                    result.points.push(
+                        new Point2D(
+                            a1.x + ua * (a2.x - a1.x),
+                            a1.y + ua * (a2.y - a1.y)
+                        )
+                    );
+                } else {
+                    result = new Intersection();
+                }
+            } else {
+                if ( ua_t === 0 || ub_t === 0 ) {
+                    result = new Intersection("Coincident");
+                } else {
+                    result = new Intersection("Parallel");
+                }
+            }
+
+            return result;
+        },
+
+
+        /**
+         *  intersectRayRay
+         *
+         *  @param {Point2D} a1
+         *  @param {Point2D} a2
+         *  @param {Point2D} b1
+         *  @param {Point2D} b2
+         *  @returns {Intersection}
+         */
+        intersectRayRay: function(a1, a2, b1, b2) {
+            var result;
+
+            var ua_t = (b2.x - b1.x) * (a1.y - b1.y) - (b2.y - b1.y) * (a1.x - b1.x);
+            var ub_t = (a2.x - a1.x) * (a1.y - b1.y) - (a2.y - a1.y) * (a1.x - b1.x);
+            var u_b  = (b2.y - b1.y) * (a2.x - a1.x) - (b2.x - b1.x) * (a2.y - a1.y);
+
+            if ( u_b !== 0 ) {
+                var ua = ua_t / u_b;
+
+                result = new Intersection();
+                result.points.push(
+                    new Point2D(
+                        a1.x + ua * (a2.x - a1.x),
+                        a1.y + ua * (a2.y - a1.y)
+                    )
+                );
+            } else {
+                if ( ua_t === 0 || ub_t === 0 ) {
+                    result = new Intersection("Coincident");
+                } else {
+                    result = new Intersection("Parallel");
+                }
+            }
+
+            return result;
+        }
+    };
+
+    for(var i = 0; i < argLength; i++) {
+        var arg = arguments[i];
+        for(var key in arg) {
+            if(arg.hasOwnProperty(key)) {
+                intersectionFunctions[key] = arg[key];
+            }
+        }
+    }
+
+    var composedShapeMethods = {};
+    composedShapeMethods[IPTYPE.PATH] = intersectionFunctions.intersectPathShape;
+    composedShapeMethods[IPTYPE.POLYLINE] = intersectionFunctions.intersectLinesShape;
+    composedShapeMethods[IPTYPE.POLYGON] = intersectionFunctions.intersectLinesShape;
+    composedShapeMethods[IPTYPE.RECT] = intersectionFunctions.intersectLinesShape;
+    composedShapeMethods[IPTYPE.ROUNDRECT] = intersectionFunctions.intersectPathShape;
+    composedShapeMethods[IPTYPE.ARC] = intersectionFunctions.intersectArcShape;
+
+
+    return intersect;
 
 }
