+++ lib/path.js
@@ -12,9 +12,7 @@ Token.prototype.init = function(type, text) {
 Token.prototype.typeis = function(type) {
 	return this.type == type;
 }
-Path.prototype = new Shape();
-Path.prototype.constructor = Path;
-Path.superclass = Shape.prototype;
+var Path = {};
 Path.COMMAND = 0;
 Path.NUMBER = 1;
 Path.EOD = 2;
